Update Debian/Ubuntu download section.
Previously, only `echo` was elevated to run as superuser, while the shell (which handled the output redirection and thus writing the sources list) ran as a normal user. This causes a problem since the sources.list.d directory is not writeable by most users.

The new version pipes the output to `tee`, which is elevated to run as root, so there is no permissions conflict.