Redesigns the homepage - no news, add images
The old homepage had news that quickly got outdated.
I removed the news section and replaced it with
the existing screenshots, hosted on Github.

I removed the header links to the outdated forum
and rearranged things so that they made a little more sense.
I also removed the "about" link, because it just linked
to the top of the page. Not much point.

Adds .gitignore for Jekyll build files

Closes #312